put pages back to grid etc
