+++ apps/next/pages/grid.tsx
@@ -1,0 +1,5 @@
+import { GridScene } from 'app/features/grid/GridScene'
+
+export default function Grid() {
+  return <GridScene />
+}
